Renamed interface to respect file naming convention
